Comment Task Complete

+++ WebApplication-API-version2/Controllers/PersonController.cs
@@ -5,6 +5,8 @@ using WebApplication_API_version2.Mappers; // Imports the mapping methods for co
 using Microsoft.EntityFrameworkCore;
 using System.Threading.Tasks;
 using WebApplication_API_version2.Interfaces;
+using WebApplication_API_version2.FilterQuery;
+
 
 
 
@@ -29,18 +31,28 @@ namespace WebApplication_API_version2.Controllers // Defines the namespace for t
 
 
         [HttpGet] // Specifies that this method responds to HTTP GET requests.
-        public async Task<IActionResult> GetAll() // Method to get all persons from the database.
+        public async Task<IActionResult> GetAll([FromQuery] QueryObject query) // Method to get all persons from the database.
         {
-            var persons = await _personRepo.GetAllAsync();
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
+            var persons = await _personRepo.GetAllAsync( query);
             var personDto=persons.Select(s => s.ToPersonDto()); // Retrieves all persons, converts them to a list, and maps them to DTOs.
             return Ok(personDto); // Returns the list of persons as a 200 OK response.
         }
 
 
 
-        [HttpGet("{Id}")] // Specifies that this method responds to GET requests with an ID parameter in the URL.
+        [HttpGet("{Id:int}")] // Specifies that this method responds to GET requests with an ID parameter in the URL.
         public async Task<IActionResult> GetById([FromRoute] int Id) // Method to get a person by their ID.
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var person =await _personRepo.GetByIdAsync(Id); // Finds the person with the given ID in the database.
             if (person == null) // Checks if the person is not found.
             {
@@ -54,6 +66,11 @@ namespace WebApplication_API_version2.Controllers // Defines the namespace for t
         [HttpPost] // Specifies that this method responds to HTTP POST requests.
         public async Task<IActionResult> Create([FromBody] CreatePersonDTO personDto) // Method to create a new person.
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var personModel = personDto.ToPersonFromCreateDto(); // Maps the incoming DTO to a Person model.
             await _personRepo.CreateAsync(personModel);
             return CreatedAtAction(nameof(GetById), new { id = personModel.Id }, personModel.ToPersonDto()); // Returns a 201 Created response, including the new person's data.
@@ -61,9 +78,14 @@ namespace WebApplication_API_version2.Controllers // Defines the namespace for t
 
 
 
-        [HttpPut("{id}")] // Specifies that this method responds to HTTP PUT requests with an ID parameter in the URL.
+        [HttpPut("{id:int}")] // Specifies that this method responds to HTTP PUT requests with an ID parameter in the URL.
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdatePersonDTO updateDto) // Method to update an existing person.
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var personModel = await _personRepo.UpdateAsync(id,updateDto); // Finds the person by ID.
             if (personModel == null) // Checks if the person is not found.
             {
@@ -74,9 +96,14 @@ namespace WebApplication_API_version2.Controllers // Defines the namespace for t
 
 
 
-        [HttpDelete("{id}")] // Specifies that this method responds to HTTP DELETE requests with an ID parameter in the URL.
+        [HttpDelete("{id:int}")] // Specifies that this method responds to HTTP DELETE requests with an ID parameter in the URL.
         public async Task<IActionResult> Delete([FromRoute] int id) // Method to delete an existing person.
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+            
             var personModel = await _personRepo.DeleteAsync(id);
             if (personModel == null) // Checks if the person is not found.
             {
